reddit-client/requirements.txt
termcolor
praw
python-dotenv
groq
